making use of cufile configurable. Also parameterized tests so that both the RAM and CuFile versions are tested.
