+++ README.md
@@ -43,6 +43,7 @@ dataset = BigWigDataset(
     batches_per_epoch=20,
     maximum_unknown_bases_fraction=0.1,
     sequence_encoder="onehot",
+    use_cufile=True,
 )
 
 for encoded_sequences, epigenetics_profiles in dataset:
