+++ bigwig_loader/bigwig.py
@@ -15,7 +15,8 @@ import pandas as pd
 from ncls import NCLS
 
 from bigwig_loader.gpu_decompressor import Decoder
-from bigwig_loader.memory_bank_cufile import CuFileMemoryBank as MemoryBank
+from bigwig_loader.memory_bank import MemoryBank
+from bigwig_loader.memory_bank import create_memory_bank
 from bigwig_loader.merge_intervals import merge_intervals
 from bigwig_loader.parser import BBIHeader
 from bigwig_loader.parser import ChromosomeTreeHeader
@@ -31,7 +32,13 @@ from bigwig_loader.util import get_standard_chromosomes
 
 
 class BigWig:
-    def __init__(self, path: Path, id: Optional[int] = None, scale: float = 1.0):
+    def __init__(
+        self,
+        path: Path,
+        id: Optional[int] = None,
+        scale: float = 1.0,
+        use_cufile: bool = True,
+    ):
         """
         Create a BigWig object representing one BigWig file.
         Args:
@@ -39,8 +46,10 @@ class BigWig:
             id: integer representing the file in a collection
                 when part of a collection.
             scale: scale values in bigwig file by this number.
+            use_cufile: whether to use kvikio cuFile to directly load data from file to
+                GPU memory.
         """
-
+        self._use_cufile = use_cufile
         with open(path, "rb") as bigwig:
             self.path = path
             self.id = id
@@ -93,6 +102,7 @@ class BigWig:
             self.path,
             chunk_sizes=self.reference_data["data_size"],  # type: ignore
             chunk_offsets=self.reference_data["data_offset"],  # type: ignore
+            use_cufile=self._use_cufile,
         )
 
     def chromosomes(
@@ -162,7 +172,7 @@ class BigWig:
 
         """
         if memory_bank is None:
-            memory_bank = MemoryBank(elastic=True)
+            memory_bank = create_memory_bank(elastic=True, use_cufile=self._use_cufile)
         if decoder is None:
             decoder = Decoder(
                 max_rows_per_chunk=self.max_rows_per_chunk,
@@ -201,7 +211,7 @@ class BigWig:
         for i in range(0, len(offsets), batch_size):
             memory_bank.reset()
             memory_bank.add_many(
-                file_handle=self.store.cufile_handle,
+                file_handle=self.store.file_handle,
                 offsets=offsets[i : i + batch_size],
                 sizes=sizes[i : i + batch_size],
                 skip_bytes=2,
