+++ bigwig_loader/store.py
@@ -9,13 +9,16 @@ from kvikio.cufile import CuFile
 class BigWigStore:
     def __init__(
         self,
-        fn: Union[str, Path],
+        path: Union[str, Path],
         chunk_offsets: npt.NDArray[np.int64],
         chunk_sizes: npt.NDArray[np.int64],
+        use_cufile: bool = True,
     ) -> None:
-        self._filename = fn
-        self._fh = open(fn, "rb", buffering=0)
-        self.cufile_handle = CuFile(fn, flags="r")
+        self.path = path
+        if use_cufile:
+            self.file_handle = CuFile(path, flags="r")
+        else:
+            self.file_handle = open(path, "rb", buffering=0)
         self.chunk_offsets = chunk_offsets
         self.chunk_sizes = chunk_sizes
