+++ tests/conftest.py
@@ -31,9 +31,14 @@ def bigwig_path():
     return get_example_bigwigs_files()
 
 
+@pytest.fixture(params=[True, False])
+def use_cufile(request):
+    return request.param
+
+
 @pytest.fixture
-def collection(bigwig_path):
-    collection = BigWigCollection(bigwig_path, first_n_files=2)
+def collection(bigwig_path, use_cufile):
+    collection = BigWigCollection(bigwig_path, first_n_files=2, use_cufile=use_cufile)
     return collection
