+++ tests/test_against_pybigwig.py
@@ -25,9 +25,9 @@ class PyBigWigCollection:
         )
 
 
-def test_same_output(bigwig_path):
+def test_same_output(bigwig_path, use_cufile):
     pybigwig_collection = PyBigWigCollection(bigwig_path, first_n_files=2)
-    collection = BigWigCollection(bigwig_path, first_n_files=2)
+    collection = BigWigCollection(bigwig_path, first_n_files=2, use_cufile=use_cufile)
 
     df = pd.read_csv(config.example_positions, sep="\t")
     df = df[df["chr"].isin(collection.get_chromosomes_present_in_all_files())]
